# Fix bug in core_number with bidirectional edges