```julia
julia> Base.Threads.nthreads()
2

julia> g = loadsnap(:amazon0302)
{262111, 1234877} directed simple UInt32 graph

julia> kcore_list = core_number(g);

julia> kcore_list = convert(Array{Int32}, kcore_list);

julia> key = unique(kcore_list)
10-element Array{Int32,1}:
  6
  7
  8
  1
 10
  4
  3
  5
  2
  9

julia> d1=Dict([(i,count(x->x==i,kcore_list)) for i in key])
Dict{Int32,Int64} with 10 entries:
  7  => 60860
  4  => 7555
  9  => 2023
  10 => 1254
  2  => 4717
  3  => 4747
  8  => 14203
  5  => 18808
  6  => 145331
  1  => 2613

julia> kcore_list = Degeneracy.core_number(g, Degeneracy.KabirMadduri(frac=0.5));

julia> kcore_list = convert(Array{Int32}, kcore_list);

julia> key = unique(kcore_list)
10-element Array{Int32,1}:
  6
  7
  8
  1
 10
  4
  3
  5
  2
  9

julia> d2=Dict([(i,count(x->x==i,kcore_list)) for i in key])
Dict{Int32,Int64} with 10 entries:
  7  => 60860
  4  => 7555
  9  => 2023
  10 => 1254
  2  => 4717
  3  => 4747
  8  => 14203
  5  => 18808
  6  => 145331
  1  => 2613

julia> d1 == d2
true
```