@sbromberger 2020-05-09
Any performance regression? I was thinking that this fix might make things faster since all_neighbors is more expensive than in_neighbors and out_neighbors combined.