@sbromberger 2020-05-09
As I suspected, the new code is faster:

OLD:
```
julia> @benchmark LightGraphs.Degeneracy.core_number(g)
BenchmarkTools.Trial: 
  memory estimate:  48.79 MiB
  allocs estimate:  814289
  --------------
  minimum time:     159.976 ms (3.15% GC)
  median time:      161.335 ms (3.12% GC)
  mean time:        164.972 ms (3.41% GC)
  maximum time:     197.475 ms (3.42% GC)
  --------------
  samples:          31
  evals/sample:     1
julia> @benchmark LightGraphs.Degeneracy.core_number(g, LightGraphs.Degeneracy.KabirMadduri())
BenchmarkTools.Trial: 
  memory estimate:  57.16 MiB
  allocs estimate:  1105577
  --------------
  minimum time:     68.858 ms (0.00% GC)
  median time:      80.235 ms (0.00% GC)
  mean time:        89.887 ms (15.16% GC)
  maximum time:     151.477 ms (44.27% GC)
  --------------
  samples:          56
  evals/sample:     1
```

THIS FIX:
```
julia> @benchmark LightGraphs.Degeneracy.core_number(g)
BenchmarkTools.Trial: 
  memory estimate:  5.00 MiB
  allocs estimate:  12
  --------------
  minimum time:     126.809 ms (0.00% GC)
  median time:      128.202 ms (0.00% GC)
  mean time:        128.317 ms (0.04% GC)
  maximum time:     131.338 ms (0.00% GC)
  --------------
  samples:          39
  evals/sample:     1
julia> @benchmark LightGraphs.Degeneracy.core_number(g, LightGraphs.Degeneracy.KabirMadduri())
BenchmarkTools.Trial: 
  memory estimate:  13.03 MiB
  allocs estimate:  279532
  --------------
  minimum time:     60.437 ms (0.00% GC)
  median time:      64.416 ms (0.00% GC)
  mean time:        65.162 ms (1.90% GC)
  maximum time:     78.181 ms (8.33% GC)
  --------------
  samples:          77
  evals/sample:     1
```